Follow 302 redirects with curl
I noticed that, after running this script, the jar file was only 144 bytes.  Running the commented-out `wget` command worked for me, so I re-ran the `curl` command with `-v`.  I noticed that the first response was a 302 redirect.  It seemed that `curl` needs an extra nudge to follow redirects.  Looking in the man page I found `-L`, which worked for me.